+++ app/Http/Controllers/api/ServicioController.php
@@ -13,13 +13,19 @@ use App\Models\Servicio;
 use App\Http\Requests\StoreServiciosRequest;
 use App\Http\Requests\UpdateServiciosRequest;   
 
+use Synfony\Componets\HttpFoundation\Response;
+
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;    
+
 class ServicioController extends Controller
 {
     public function index(){
-        $servicio = Servicio::all();
+        $this->authorize ('Ver servicios');
+        $servicio = Servicio::with(['users', 'membresias'])->get();
         return new ServicioCollection($servicio);
     }
     public function show($id){
+        $this ->authorize('Ver servicios');
         $servicio = Servicio::find($id);
         if ($servicio) {
             return new ServicioResource($servicio);
@@ -28,6 +34,7 @@ class ServicioController extends Controller
         }
     }
     public function store(StoreMembresiasRequest $request){
+        $this->authorize('Crear servicios');
         $servicio= Servicio::create ($request->validate());
 
         return (new ServicioResource($servicio))
@@ -35,6 +42,7 @@ class ServicioController extends Controller
         ->setStatusCode(201);
     }
     public function update(StoreMembresiasRequest $request, $id){
+        $this->authorize('Actualizar servicios');
         $servicio = Servicio::find($id);
         if ($servicio) {                
             $servicio::update($request->validate());
@@ -45,6 +53,7 @@ class ServicioController extends Controller
     }
     
     Public function destroy($id){
+        $this->authorize('Eliminar servicios');
         $servicio = Servicio::find($id);
         if ($servicio) {
             $servicio->delete();
